+++ Projetos/FrontEnd/front-processos/src/App.js
@@ -1,8 +1,10 @@
 import { ThemeProvider } from '@material-ui/core'
 import { getTheme } from './theme/Theme'
 import { useState } from 'react'
-import { SearchBar } from './components/SearchBar/SearchBar'
-import { HomeContainer } from './components/layout/HomeContainer/HomeContainer'
+import { Navbar } from './components/Navbar'
+import { ReactKeycloakProvider } from '@react-keycloak/web'
+import keycloak from './keycloak'
+import { AppRouter } from './routes'
 
 const App = () => {
   const [darkMode, setDarkMode] = useState(false)
@@ -13,11 +15,21 @@ const App = () => {
 
   const theme = getTheme(darkMode)
 
+  const eventLogger = (event, error) => {
+    console.log('onKeycloakEvent', event, error)
+  }
+
+  const tokenLogger = (tokens) => {
+    console.log('onKeycloakTokens', tokens)
+  }
+
   return (
-    <ThemeProvider theme={theme}>
-      <HomeContainer toggleTheme={toggleTheme} />
-      <SearchBar />
-    </ThemeProvider>
+    <ReactKeycloakProvider authClient={keycloak} onEvent={eventLogger} onTokens={tokenLogger}>
+      <ThemeProvider theme={theme}>
+        <Navbar toggleTheme={toggleTheme} />
+        <AppRouter />
+      </ThemeProvider>
+    </ReactKeycloakProvider>
   )
 }
