+++ Projetos/FrontEnd/front-processos/src/pages/ListagemProcesso.js
@@ -1,0 +1,4 @@
+const ListagemProcesso = () => {
+  return <>Listagem de processo</>
+}
+export default ListagemProcesso
